pyproject.toml
[tool.poetry]
name = "python-aws"
version = "0.1.0"
description = ""
authors = ["Timileyin <sadiquetimileyin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11, <3.13"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
pandas = "^2.2.1"
pyarrow = "^15.0.0"
s3fs = "^2024.3.1"
apache-airflow = "^2.9.0"
pytest = "^8.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
